Apply review suggestions
Remove standalone_event_pool in favor of a context-owned event pool cache
Update urEventCreateExp implementation by borrowing and allocating from the cache
Move event_descriptor back into event_pool_cache as a private nested struct
Drop event_descriptor's unused equality operator and hash struct
