+++ unified-runtime/test/adapters/level_zero/v2/CMakeLists.txt
@@ -57,7 +57,6 @@ add_l0_v2_devices_test(event_pool
     ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/event_provider_counter.cpp
     ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/event_provider_normal.cpp
     ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/event.cpp
-    ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/standalone_event_pool.cpp
     ${PROJECT_SOURCE_DIR}/source/ur/ur.cpp
 )
 
@@ -79,7 +78,6 @@ add_l0_v2_devices_test(batched_queue
     ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/event_provider_counter.cpp
     ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/event_provider_normal.cpp
     ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/event.cpp
-    ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/standalone_event_pool.cpp
     ${PROJECT_SOURCE_DIR}/source/ur/ur.cpp
      ${PROJECT_SOURCE_DIR}/source/adapters/level_zero/v2/command_list_cache.cpp
 )
